add login

+++ crud.py
@@ -2,6 +2,7 @@ import json
 import time
 
 import click
+from flask_login import login_required, current_user
 from flask_restful import Resource, reqparse
 from sqlalchemy.exc import IntegrityError
 
@@ -10,12 +11,24 @@ from ext import db, milvus
 from models import Question, QuestionSet
 
 
+def can_maintain(user, question_set):
+    if user == question_set.owner:
+        return True
+    if user in question_set.maintainer:
+        return True
+    return False
+
+
 class QuestionAPI(Resource):
+    decorators = [login_required]
 
     # GET 显示数据并不全，很多字段都没显示
     # 或许 GET 加参数来决定需不需要某些数据？
     # 问题库包含的问题，数据量可能比较大
     # 但是用户可能只需要问题库名字？
+
+    # Question 相关没有判断user权限
+    # QuestionSet 的 append 也应该判断是否允许加入
     def get(self, qid):
         question = Question.query.get(qid)
         if question:
@@ -36,6 +49,7 @@ class QuestionAPI(Resource):
             question.title = title
             question.content = content
             question.embedding = json.dumps(embedding)
+            question.modified_by_id = current_user.id
             db.session.commit()
             for sid in question.belongs.with_entities(QuestionSet.id).all():
                 collection_name = '_' + str(sid[0])
@@ -57,6 +71,7 @@ class QuestionAPI(Resource):
 
 
 class QuestionGroupAPI(Resource):
+    decorators = [login_required]
 
     def get(self):  # 批量获取信息
         pass
@@ -75,6 +90,7 @@ class QuestionGroupAPI(Resource):
             title, content = titles[0], contents[0]
             embedding = json.dumps(rocketqa.get_para(title, content))  # dumps 只出现在了这里
             question = Question(title=title, content=content, embedding=embedding)
+            question.created_by_id = current_user.id
             db.session.add(question)
             db.session.commit()
             return {'message': '问题创建成功', 'id': question.id}
@@ -83,6 +99,7 @@ class QuestionGroupAPI(Resource):
             questions = []
             for title, content, embedding in zip(titles, contents, embeddings):
                 question = Question(title=title, content=content, embedding=json.dumps(embedding))
+                question.created_by_id = current_user.id
                 questions.append(question)
             db.session.add_all(questions)
             db.session.commit()
@@ -91,10 +108,13 @@ class QuestionGroupAPI(Resource):
 
 
 class QuestionSetAPI(Resource):
+    decorators = [login_required]
 
     def get(self, sid):
         question_set = QuestionSet.query.get(sid)
         if question_set:
+            if not can_maintain(current_user, question_set):
+                return {'message': '没权限'}, 400
             return {'id': sid,
                     'name': question_set.name,
                     'questions': [qid[0] for qid in question_set.questions.with_entities(Question.id).all()]
@@ -116,12 +136,16 @@ class QuestionSetAPI(Resource):
         if qs is None:
             return {'message': '问题库不存在'}, 400
 
+        if not can_maintain(current_user, qs):
+            return {'message': '没权限'}, 400
+
         if op == 'append':
             if not qids:
                 return {'message': '问题ID不能为空'}, 400
             start = time.time()
             questions = Question.query.filter(Question.id.in_(qids)).all()
             qs.questions.extend(questions)
+            qs.modified_by_id = current_user.id
             try:
                 db.session.commit()
             except Exception as e:
@@ -144,6 +168,7 @@ class QuestionSetAPI(Resource):
             questions = Question.query.filter(Question.id.in_(qids)).all()
             for question in questions:
                 qs.questions.remove(question)
+            qs.modified_by_id = current_user.id
             try:
                 db.session.commit()
             except Exception as e:
@@ -157,14 +182,20 @@ class QuestionSetAPI(Resource):
             if not name:
                 return {'message': '名称不能为空'}, 400
             qs.name = name
+            qs.modified_by_id = current_user.id
             db.session.commit()
             return {'message': '问题库更名'}
+
+        # add maintainer
+        # change owner
 
         return {'message': '无法理解的操作'}, 400
 
     def delete(self, sid):
         qs = QuestionSet.query.get(sid)
         if qs:
+            if not can_maintain(current_user, qs):
+                return {'message': '没权限'}, 400
             milvus.drop_collection('_' + str(qs.id))
             db.session.delete(qs)
             db.session.commit()
@@ -173,6 +204,7 @@ class QuestionSetAPI(Resource):
 
 
 class QuestionSetGroupAPI(Resource):
+    decorators = [login_required]
 
     def post(self):
         args = reqparse.RequestParser() \
@@ -184,22 +216,13 @@ class QuestionSetGroupAPI(Resource):
         # qids = args['questions']
 
         qs = QuestionSet(name=name)
-        # embeddings = []
-        # if qids:
-        #     questions = Question.query.fliter(Question.id.in_(qids)).all()
-        #     if len(questions) != len(qids):
-        #         return {'message': '检查问题'}, 400
-        #     qs.questions.extend(questions)
-        #     embeddings = [json.loads(question.embedding) for question in questions]
-        # for qid in qids:
-        #     question = Question.query.get(qid)
-        #     qs.questions.append(question)
-        #     embeddings.append(json.loads(question.embedding))
+        qs.created_by_id = current_user.id
+        qs.owner_id = current_user.id
+
         db.session.add(qs)
         db.session.commit()
 
         collection_name = '_' + str(qs.id)
         milvus.create_collection(collection_name)  # 按理说这个名字的 collection 是不存在的
-        # if len(embeddings) > 0:
-        #     milvus.insert(collection_name, embeddings, qids)
+
         return {'message': '问题库创建成功'}
